+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Rename restart files in GCHP integration tests (as we do in non-test runs)
 - The `integrationTestCreate.sh` copy `integrationDiffTest.sh` to the integration test scripts folder
 - Request 6 hours of execution time for GEOS-Chem Classic integration tests
+- Invert directory structure where integration and parallel test scripts are stored
 
 ## [Unreleased 14.2.0]
 ### Added